fix(implemented freds pr changes): implemented freds pr changes
