+++ interfaces/index.ts
@@ -56,7 +56,3 @@ export type Organization = {
   id: number;
   name: string;
 };
-
-export type OrganizationApplication = {
-  id: number;
-};
