+++ pages/api/organization-applications/index.ts
@@ -2,7 +2,7 @@ import { PrismaClient, OrganizationApplication } from '@prisma/client';
 import Joi, { ValidationError } from 'joi';
 import { NextApiRequest, NextApiResponse } from 'next';
 import CreateError, { MethodNotAllowed } from 'utils/error';
-import OrganizationApplicationSchema from '../../../interfaces/organization_application';
+import OrganizationApplicationSchema from '../../../interfaces/organizationApplication';
 
 const prisma = new PrismaClient();
 
@@ -15,26 +15,7 @@ export const createOrganizationApp = async (
   }
   const data = value as OrganizationApplication;
   const createdOrgApp = await prisma.organizationApplication.create({
-    data: {
-      applicationStatus: data.applicationStatus,
-      organizationName: data.organizationName,
-      contactName: data.contactName,
-      contactEmail: data.contactEmail,
-      organizationType: data.organizationType,
-      workType: data.workType,
-      address: data.address,
-      lat: data.lat,
-      long: data.long,
-      missionStatement: data.missionStatement,
-      shortHistory: data.shortHistory,
-      keyValue: data.keyValue,
-      lgbtqDemographic: data.lgbtqDemographic,
-      raceDemographic: data.raceDemographic,
-      ageDemographic: data.ageDemographic,
-      capacity: data.capacity,
-      ein: data.ein,
-      foundingDate: data.foundingDate,
-    },
+    data,
   });
   return createdOrgApp;
 };
@@ -43,11 +24,6 @@ export default async (
   req: NextApiRequest,
   res: NextApiResponse
 ): Promise<void> => {
-  const orgAppId = req.query.id as string;
-  if (Joi.number().validate(orgAppId).error) {
-    return CreateError(400, `ID ${orgAppId} is not a number`, res);
-  }
-
   if (req.method !== 'POST') {
     return MethodNotAllowed(req.method, res);
   }
@@ -58,10 +34,9 @@ export default async (
     }
     return res.json(orgapp);
   } catch (err) {
-    return CreateError(
-      500,
-      `Failed to create organization application ${orgAppId}`,
-      res
-    );
+    if (err instanceof ValidationError) {
+      return CreateError(400, err.message, res);
+    }
+    return CreateError(500, `Failed to create organization application.`, res);
   }
 };
